Reformatting and mirror testing.
