+++ src/main/java/org/spoutcraft/launcher/GUI/HyperlinkJLabel.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.spoutcraft.launcher.GUI;
+package org.spoutcraft.launcher.gui;
 
 import java.awt.Desktop;
 import java.awt.event.MouseEvent;
