+++ src/main/java/org/spoutcraft/launcher/GUI/UpdateDialog.java
@@ -1,4 +1,4 @@
-package org.spoutcraft.launcher.GUI;
+package org.spoutcraft.launcher.gui;
 
 import java.awt.BorderLayout;
 import java.awt.FlowLayout;
