+++ src/main/java/org/spoutcraft/launcher/async/Download.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.spoutcraft.launcher.AsyncDownload;
+package org.spoutcraft.launcher.async;
 
 import java.io.*;
 import java.net.*;
