+++ src/main/java/org/spoutcraft/launcher/exception/OutdatedMCLauncherException.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.spoutcraft.launcher.Exceptions;
+package org.spoutcraft.launcher.exception;
 
 public class OutdatedMCLauncherException extends Exception{
