+++ src/main/java/org/spoutcraft/launcher/exception/UnsupportedOSException.java
@@ -14,36 +14,35 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.spoutcraft.launcher.Logging;
-import java.io.InvalidObjectException;
-import java.io.ObjectStreamException;
-import java.util.logging.Level;
+package org.spoutcraft.launcher.exception;
 
-public class SystemListenerLevel extends Level {
+public class UnsupportedOSException extends Exception{
 	
-
 	/**
 	 * 
 	 */
-	private static final long serialVersionUID = 1L;
-
-
-	private SystemListenerLevel(String name, int value) {
-		super(name, value);
+	private static final long serialVersionUID = 521197180311147444L;
+	private final Throwable cause;
+	private final String message;
+	
+	public UnsupportedOSException(String message) {
+	  this(null, message);
 	}
 
-	public static Level STDOUT = new SystemListenerLevel("STDOUT", Level.INFO.intValue()+53);
+	public UnsupportedOSException(Throwable throwable, String message) {
+	  this.cause = null;
+	  this.message = message;
+	}
 
-	public static Level STDERR = new SystemListenerLevel("STDERR", Level.INFO.intValue()+54);
+	public UnsupportedOSException() {
+	  this(null, "Bad login");
+	}
 
+	public Throwable getCause() {
+	  return this.cause;
+	}
 
-	protected Object readResolve()
-	throws ObjectStreamException {
-		if (this.intValue() == STDOUT.intValue())
-			return STDOUT;
-		if (this.intValue() == STDERR.intValue())
-			return STDERR;
-		throw new InvalidObjectException("Unknown instance :" + this);
-	}		
-	
+	public String getMessage() {
+	  return this.message;
+	}
 }
