+++ src/main/java/org/spoutcraft/launcher/logs/SystemListenerLevel.java
@@ -14,35 +14,36 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.spoutcraft.launcher.Exceptions;
+package org.spoutcraft.launcher.logs;
+import java.io.InvalidObjectException;
+import java.io.ObjectStreamException;
+import java.util.logging.Level;
 
-public class UnsupportedOSException extends Exception{
+public class SystemListenerLevel extends Level {
 	
+
 	/**
 	 * 
 	 */
-	private static final long serialVersionUID = 521197180311147444L;
-	private final Throwable cause;
-	private final String message;
+	private static final long serialVersionUID = 1L;
+
+
+	private SystemListenerLevel(String name, int value) {
+		super(name, value);
+	}
+
+	public static Level STDOUT = new SystemListenerLevel("STDOUT", Level.INFO.intValue()+53);
+
+	public static Level STDERR = new SystemListenerLevel("STDERR", Level.INFO.intValue()+54);
+
+
+	protected Object readResolve()
+	throws ObjectStreamException {
+		if (this.intValue() == STDOUT.intValue())
+			return STDOUT;
+		if (this.intValue() == STDERR.intValue())
+			return STDERR;
+		throw new InvalidObjectException("Unknown instance :" + this);
+	}		
 	
-	public UnsupportedOSException(String message) {
-	  this(null, message);
-	}
-
-	public UnsupportedOSException(Throwable throwable, String message) {
-	  this.cause = null;
-	  this.message = message;
-	}
-
-	public UnsupportedOSException() {
-	  this(null, "Bad login");
-	}
-
-	public Throwable getCause() {
-	  return this.cause;
-	}
-
-	public String getMessage() {
-	  return this.message;
-	}
 }
